feat: CNN network created and trained
The CNN consists of 4 layers using dropout. The fourth layer is a dense layer.
To get satisfactory results the images are scaled by dividing each RGB channel 
by 255 in the path_to_tensor function.

The accuracy metric was used but it needs to be changed because the 0 and 1 
classes are not balanced. Alternatively image augmentation can be applied.